Fix 'Add Tag to Selected' not emitting badge signals
