+++ src/tagstudio/qt/ts_qt.py
@@ -177,6 +177,9 @@ class QtDriver(DriverMixin, QObject):
 
     SIGTERM = Signal()
 
+    favorite_updated = Signal(bool)
+    archived_updated = Signal(bool)
+
     tag_manager_panel: PanelModal | None = None
     color_manager_panel: TagColorManager | None = None
     ignore_modal: PanelModal | None = None
@@ -560,12 +563,12 @@ class QtDriver(DriverMixin, QObject):
 
         self.main_window.search_field.textChanged.connect(self.update_completions_list)
 
-        self.main_window.preview_panel.field_containers_widget.archived_updated.connect(
+        self.archived_updated.connect(
             lambda hidden: self.update_badges(
                 {BadgeType.ARCHIVED: hidden}, origin_id=0, add_tags=False
             )
         )
-        self.main_window.preview_panel.field_containers_widget.favorite_updated.connect(
+        self.favorite_updated.connect(
             lambda hidden: self.update_badges(
                 {BadgeType.FAVORITE: hidden}, origin_id=0, add_tags=False
             )
@@ -801,6 +804,19 @@ class QtDriver(DriverMixin, QObject):
             )
         )
 
+    def emit_badge_signals(self, tag_ids: list[int] | set[int], emit_on_absent: bool = True):
+        """Emit any connected signals for updating badge icons."""
+        logger.info("[emit_badge_signals] Emitting", tag_ids=tag_ids, emit_on_absent=emit_on_absent)
+        if TAG_ARCHIVED in tag_ids:
+            self.archived_updated.emit(True)  # noqa: FBT003
+        elif emit_on_absent:
+            self.archived_updated.emit(False)  # noqa: FBT003
+
+        if TAG_FAVORITE in tag_ids:
+            self.favorite_updated.emit(True)  # noqa: FBT003
+        elif emit_on_absent:
+            self.favorite_updated.emit(False)  # noqa: FBT003
+
     def add_tag_action_callback(self):
         panel = BuildTagPanel(self.lib)
         self.modal = PanelModal(
@@ -852,6 +868,7 @@ class QtDriver(DriverMixin, QObject):
         selected: list[int] = self.selected
         self.main_window.thumb_layout.add_tags(selected, tag_ids)
         self.lib.add_tags_to_entries(selected, tag_ids)
+        self.emit_badge_signals(tag_ids)
 
     def delete_files_callback(self, origin_path: str | Path, origin_id: int | None = None):
         """Callback to send on or more files to the system trash.
